feat: 간단한 게시판 CRUD 기능 구현

+++ src/main/java/JHboard/project/domain/board/service/BoardServiceImpl.java
@@ -1,6 +1,7 @@
 package JHboard.project.domain.board.service;
 
 
+import JHboard.project.domain.board.dto.BoardRqDto;
 import JHboard.project.domain.board.entity.Board;
 import JHboard.project.domain.board.repository.BoardRepository;
 import java.util.List;
@@ -36,6 +37,20 @@ public class BoardServiceImpl implements BoardService{
   @Override
   @Transactional
   public void delete(Long boardId) {
+    Optional<Board> board = boardRepository.findById(boardId);
+    if(board.isEmpty()){
+      throw new RuntimeException("존재하지 않은 게시글입니다.");
+    }
     boardRepository.deleteById(boardId);
+  }
+
+  @Override
+  @Transactional
+  public void updateBoard(Long boardId, BoardRqDto boardRqDto) {
+    Optional<Board> board = boardRepository.findById(boardId);
+    if(board.isEmpty()){
+      throw new RuntimeException("존재하지 않는 게시글입니다.");
+    }
+    board.get().updateBoard(boardRqDto);
   }
 }
